+++ src/Console/ExtendSeeder.php
@@ -2,11 +2,12 @@
 
 namespace Touhidurabir\SeedExtender\Console;
 
-use Throwable;
 use Exception;
+use Throwable;
 use Illuminate\Console\Command;
 use Touhidurabir\StubGenerator\StubGenerator;
 use Touhidurabir\SeedExtender\Concerns\SchemaHelper;
+use Touhidurabir\SeedExtender\Concerns\SeederValidator;
 use Touhidurabir\SeedExtender\Exceptions\SeedExtenderSchemaException;
 use Touhidurabir\SeedExtender\Console\Concerns\CommandExceptionHandler;
 
@@ -19,6 +20,8 @@ class ExtendSeeder extends Command {
 
     use SchemaHelper;
 
+    use SeederValidator;
+
 
     /**
      * The name and signature of the console command.
@@ -28,10 +31,12 @@ class ExtendSeeder extends Command {
     protected $signature = 'make:extend-seeder
                             {class                                  : Seeder class name}
                             {--table=                               : Seeder table name}
+                            {--model=                               : will run the seeding process through given eloquent model class}
                             {--ignorables=                          : Column list to ignore}
                             {--useables=                            : Column list to use}
                             {--timestampables=created_at,updated_at : The columns related to table timestamp}
                             {--no-timestamp                         : Should allow auto merge of timestamp column value of created_at and updated_at}
+                            {--with-events                          : Will fire the eloquent model events when seeding through model class}
                             {--replace                              : Should replace existing class files}
                             {--strict                               : should do a strict checking of provided table and columns existance}';
 
@@ -100,12 +105,14 @@ class ExtendSeeder extends Command {
                                 ->to('/database/seeders/', true)
                                 ->as($this->argument('class'))
                                 ->withReplacers([
-                                    'class'         => $this->argument('class'),
-                                    'table'         => $this->option('table'),
-                                    // 'columns'       => $this->option('table') ? $this->columnList($this->option('table')) : [],
-                                    'ignorables'    => $this->option('ignorables') ? array_map('trim', explode(',', $this->option('ignorables'))) : [],
-                                    'useables'      => $this->option('useables') ? array_map('trim', explode(',', $this->option('useables'))) : [],
-                                    'timestamp'     => $this->option('no-timestamp') ? false : $this->hasColumns($this->option('table'), array_map('trim', explode(',', $this->option('timestampables')))),
+                                    'class'             => $this->argument('class'),
+                                    'table'             => $this->option('table'),
+                                    // 'columns'           => $this->option('table') ? $this->columnList($this->option('table')) : [],
+                                    'ignorables'        => $this->option('ignorables') ? array_map('trim', explode(',', $this->option('ignorables'))) : [],
+                                    'useables'          => $this->option('useables') ? array_map('trim', explode(',', $this->option('useables'))) : [],
+                                    'timestamp'         => $this->option('no-timestamp') ? false : $this->hasColumns($this->option('table'), array_map('trim', explode(',', $this->option('timestampables')))),
+                                    'seedThroughModel'  => $this->throughModelStubber(),
+
                                 ])
                                 ->replace($this->option('replace'))
                                 ->save();
@@ -139,6 +146,28 @@ class ExtendSeeder extends Command {
 
 
     /**
+     * Generate form stub segment for model based seeding and passed as string 
+     *
+     * @return string
+     */
+    private function throughModelStubber() {
+        
+        if ( ! $this->option('model') ) {
+
+            return '';
+        }
+
+        return (new StubGenerator)
+                ->from($this->generateFullPathOfStubFile('/stubs/seed_through_model.stub'), true)
+                ->withReplacers([
+                    'model'     => $this->option('model') . '::class',
+                    'quietly'   => $this->option('with-events') ? false : true,
+                ])
+                ->toString();
+    }
+
+
+    /**
      * Handle strict chekcing of passed options
      *
      * @param  bool $strict
@@ -153,49 +182,23 @@ class ExtendSeeder extends Command {
             return;
         }
 
-        if ( ! $this->option('table') ) {
-
-            throw SeedExtenderSchemaException::noTableGiven();
-        }
-
-        if ( !$this->tableExists($this->option('table')) ) {
-
-            throw SeedExtenderSchemaException::tableNotFound($this->option('table'));
-        }
-
+        $this->validateTableHasProvided($this->option('table'));
+        $this->validateGivenTableExists($this->option('table'));
+        
         collect($this->columnableOptions)
-            ->each( fn($option, $index) => $this->validateGivenColumnsList($option) );
+            ->each( function($option, $index) {
 
+                if ( $this->option($option) ) {
+
+                    $this->validateProvidedColumnsDefinedInTable(
+                        $this->option('table'),
+                        array_map('trim', explode(',', $this->option($option))),
+                        $option
+                    );
+                }
+            });
+
+        $this->validateGivenModel($this->option('model'));
     }
-
-
-    /**
-     * Validate the give columns lists 
-     *
-     * @param  string $option
-     * @return void
-     * 
-     * @throws Exception
-     */
-    private function validateGivenColumnsList(string $option = null) {
-
-        if ( ! $this->option($option) ) {
-
-            return;
-        }
-
-        $columns = array_map('trim', explode(',', $this->option($option)));
-
-        if ( $this->hasColumns($this->option('table'), $columns) ) {
-
-            return;
-        }
-
-        throw SeedExtenderSchemaException::tableColumnsNotDefined(
-            $this->option('table'), 
-            $this->arbitraryColumns($this->option('table'), $columns), 
-            "{$option} option"
-        );
-    }   
 
 }