+++ CHANGELOG.md
@@ -18,3 +18,11 @@ All notable changes to this project will be documented in this file.
 - Added the columns list handling process more dynamic.
 - Feature optimization. 
 - README update.
+
+## [1.1.0] - 2021-12-02
+- Added functionality to handle seeding via table model class.
+- Code Optimization.
+- Existing feature optimization. 
+- Tests update.
+- Bug Fix.
+- README update.